backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.10
python-jose==3.3.0
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
pydantic==2.8.0
pydantic-settings==2.2.0
python-dotenv==1.0.0
Pillow==10.4.0
aiofiles==23.2.1
email-validator==2.1.0
